자동이체 출력 코드 수정
