+++ src/component/templates/Debits.tsx
@@ -5,18 +5,11 @@ import { usePopup } from "../../hooks/popup/usePopup";
 import TotalExpense from "../molecules/TotalExpense";
 import DebitList from "../organisms/list/DebitList";
 import Register from "../molecules/Register";
-import { useContext } from "react";
-import { DataContext } from "../../App";
-import { useDebit } from "../../hooks/useDebit";
-import { useRecoilValue } from "recoil";
-import { accountState } from "../../recoil/states/ClickedState";
-import { findById } from "../../utils/FuncSet";
+import { useGetAccount } from "../../hooks/useGetAccounts";
 
 const Debits: React.FC = () => {
   const [isPop, openPop, closePop, togglePop] = usePopup();
-  const debits = useDebit();
-  const accounts = useContext(DataContext);
-  const current = useRecoilValue(accountState);
+  const accounts = useGetAccount();
 
   const toggle: VoidtoVoid = () => {
     togglePop();
@@ -33,8 +26,8 @@ const Debits: React.FC = () => {
   return (
     <LayoutVerticalFlex option={{ justifyContent: "flex-start" }}>
       <Register modalType="Debit" toggle={toggle} />
-      <DebitList debits={debits} />
-      <TotalExpense account={findById(accounts, current.id)} />
+      <DebitList debits={accounts?.debits} />
+      <TotalExpense expense={accounts?.expense} />
       {isPop && (
         <RegisterModal register={register} cancel={cancel} modalType="Debit" />
       )}
