+++ src/utils/FuncSet.ts
@@ -1,6 +1,8 @@
 import { Account } from "../classes/types/DataTypes";
 
-export const findById = (list: Account[], id: number) => {
+type FindFuncType = (list: Account[], id: number) => Account;
+
+export const findById: FindFuncType = (list: Account[], id: number) => {
   return id
     ? list.filter((account) => Number.parseInt(account.id) === id)[0]
     : list[0];
